Lock share actions during vault move/delete

+++ packages/pass/store/sagas/vaults/vault-delete.saga.ts
@@ -4,12 +4,14 @@ import { PassCrypto } from '@proton/pass/lib/crypto';
 import { deleteVault } from '@proton/pass/lib/vaults/vault.requests';
 import {
     getUserAccessIntent,
+    lockShare,
+    unlockShare,
     vaultDeleteFailure,
     vaultDeleteIntent,
     vaultDeleteSuccess,
 } from '@proton/pass/store/actions';
 import { withRevalidate } from '@proton/pass/store/request/enhancers';
-import { selectUserDefaultShareID } from '@proton/pass/store/selectors';
+import { isShareLocked, selectUserDefaultShareID } from '@proton/pass/store/selectors';
 import type { RootSagaOptions } from '@proton/pass/store/types';
 import type { Maybe } from '@proton/pass/types';
 
@@ -18,6 +20,10 @@ function* deleteVaultWorker(
     { payload: { shareId, content }, meta }: ReturnType<typeof vaultDeleteIntent>
 ) {
     try {
+        const shareLocked: boolean = yield select(isShareLocked(shareId));
+        if (shareLocked) throw new Error();
+
+        yield put(lockShare(shareId));
         yield deleteVault(shareId);
         PassCrypto.removeShare(shareId);
 
@@ -31,6 +37,8 @@ function* deleteVaultWorker(
         onItemsUpdated?.();
     } catch (e) {
         yield put(vaultDeleteFailure(meta.request.id, { shareId, content }, e));
+    } finally {
+        yield put(unlockShare(shareId));
     }
 }
